Change relu to clamp for grouped conv fwd instances
