--- library/src/tensor_operation_instance/gpu/grouped_conv3d_fwd_bias_relu/xdl/device_grouped_conv3d_fwd_bias_relu_xdl_ndhwgc_gkzyxc_ndhwgk_bf16_16x16_instance.cpp
@@ -1,58 +1,0 @@
-// SPDX-License-Identifier: MIT
-// Copyright (c) 2025, Advanced Micro Devices, Inc. All rights reserved.
-
-#include "ck/library/tensor_operation_instance/gpu/grouped_conv_fwd/device_grouped_conv_fwd_xdl_instance.hpp"
-#include "ck/library/tensor_operation_instance/add_device_operation_instance.hpp"
-
-namespace ck {
-namespace tensor_operation {
-namespace device {
-namespace instance {
-
-void add_device_grouped_conv3d_fwd_bias_relu_xdl_ndhwgc_gkzyxc_ndhwgk_bf16_16x16_instances(
-    std::vector<std::unique_ptr<DeviceGroupedConvFwdMultipleABD<3,
-                                                                NDHWGC,
-                                                                GKZYXC,
-                                                                Tuple<NDHWGK>,
-                                                                NDHWGK,
-                                                                BF16,
-                                                                BF16,
-                                                                Tuple<BF16>,
-                                                                BF16,
-                                                                PassThrough,
-                                                                PassThrough,
-                                                                AddRelu>>>& instances)
-{
-    add_device_operation_instances(instances,
-                                   device_grouped_conv_fwd_xdl_bf16_16x16_instances<3,
-                                                                                    NDHWGC,
-                                                                                    GKZYXC,
-                                                                                    Tuple<NDHWGK>,
-                                                                                    NDHWGK,
-                                                                                    ConvFwdDefault,
-                                                                                    Tuple<BF16>,
-                                                                                    AddRelu>{});
-    add_device_operation_instances(instances,
-                                   device_grouped_conv_fwd_xdl_bf16_16x16_instances<3,
-                                                                                    NDHWGC,
-                                                                                    GKZYXC,
-                                                                                    Tuple<NDHWGK>,
-                                                                                    NDHWGK,
-                                                                                    ConvFwd1x1P0,
-                                                                                    Tuple<BF16>,
-                                                                                    AddRelu>{});
-    add_device_operation_instances(instances,
-                                   device_grouped_conv_fwd_xdl_bf16_16x16_instances<3,
-                                                                                    NDHWGC,
-                                                                                    GKZYXC,
-                                                                                    Tuple<NDHWGK>,
-                                                                                    NDHWGK,
-                                                                                    ConvFwd1x1S1P0,
-                                                                                    Tuple<BF16>,
-                                                                                    AddRelu>{});
-}
-
-} // namespace instance
-} // namespace device
-} // namespace tensor_operation
-} // namespace ck
