--- test/grouped_convnd_fwd_bias_relu/CMakeLists.txt
@@ -1,4 +1,0 @@
-if(GPU_TARGETS MATCHES "gfx9")
-    add_gtest_executable(test_grouped_convnd_fwd_bias_relu test_grouped_convnd_fwd_bias_relu.cpp)
-    target_link_libraries(test_grouped_convnd_fwd_bias_relu PRIVATE utility device_grouped_conv2d_fwd_bias_relu_instance device_grouped_conv3d_fwd_bias_relu_instance)
-endif()
